+++ Observer/ProductSave.php
@@ -69,12 +69,8 @@ class ProductSave implements ObserverInterface
     /**
      * ProductSave constructor.
      * @param \Autocompleteplus\Autosuggest\Helper\Data $helper
-     * @param \Magento\ConfigurableProduct\Model\Product\Type\Configurable $configurable
      * @param \Psr\Log\LoggerInterface $logger
      * @param \Magento\Framework\Stdlib\DateTime\DateTime $date
-     * @param \Autocompleteplus\Autosuggest\Model\ResourceModel\Batch\CollectionFactory $batchCollectionFactory
-     * @param \Magento\Catalog\Model\Product $productModel
-     * @param \Autocompleteplus\Autosuggest\Model\Batch $batchModel
      */
     public function __construct(
         \Autocompleteplus\Autosuggest\Helper\Batches $helper,